Move random functions out of model
Allows it to be more easily used by other modules.
